LU-2074 build: fix 'copy into fixed size buffer' errors
Fix 'copy into fixed size buffer' defects found by Coverity
version 6.0.3:
Copy into fixed size buffer (STRING_OVERFLOW)
The fixed-size string might be overrun by copying without
checking the length.

+++ lnet/utils/debug.c
@@ -584,6 +584,7 @@ int jt_dbg_debug_kernel(int argc, char **argv)
         int fdin;
         int fdout;
         int rc;
+	int cplen = 0;
 
         if (argc > 3) {
                 fprintf(stderr, "usage: %s [file] [raw]\n", argv[0]);
@@ -600,11 +601,17 @@ int jt_dbg_debug_kernel(int argc, char **argv)
         /* If we are dumping raw (which means no conversion step to ASCII)
          * then dump directly to any supplied filename, otherwise this is
          * just a temp file and we dump to the real file at convert time. */
-        if (argc > 1 && raw)
-                strcpy(filename, argv[1]);
-        else
-                sprintf(filename, "%s"CFS_TIME_T".%u",
-			LIBCFS_DEBUG_FILE_PATH_DEFAULT, time(NULL), getpid());
+	if (argc > 1 && raw) {
+		cplen = strlcpy(filename, argv[1], sizeof(filename));
+	} else
+		cplen = snprintf(filename, sizeof(filename),
+				 "%s"CFS_TIME_T".%u",
+				 LIBCFS_DEBUG_FILE_PATH_DEFAULT, time(NULL),
+				 getpid());
+	if (cplen >= sizeof(filename)) {
+		fprintf(stderr, "File name too long: %s\n", argv[1]);
+		return 1;
+	}
 
         if (stat(filename, &st) == 0 && S_ISREG(st.st_mode))
                 unlink(filename);
